Fixed the pyrender method

+++ visualize_smpl.py
@@ -42,7 +42,7 @@ def prepare_out_dirs(prefix:str='output/', dataDirs=['img']):
     result = []
     for dataDir in dataDirs:
         output_dir = os.path.join(prefix, dataDir)
-        if os.path.exists(output_dir) and os.path.isdir(output_dir):
+        if os.path.exists(output_dir) and os.path.isdir(output_dir) and DELETE_OLD_OUTPUT:
             shutil.rmtree(output_dir)
         print('Created', output_dir)
         os.makedirs(output_dir, exist_ok=True)
@@ -96,7 +96,7 @@ def create_mesh(vertices, faces, colors=None, **kwargs):
 
 
 # project a 3D point in the world to the 2D views
-def project_to_views(points, frame_id):
+def project_to_views(meshes, frame_id):
     output_dir = prepare_out_dirs(dataDirs=['vertices_on_img'])[0]
     for cam in CAMERAS[:]:
         file_id = str(frame_id).zfill(10)
@@ -109,12 +109,17 @@ def project_to_views(points, frame_id):
         kinect_offset = np.array([0.5, 0.5])
         if color is None:
             print("File not found: ", fpath)
-        points2d = project_pose(points, params)
-        for loc2d in points2d:
-            # x, y = np.int16(np.round(loc2d - kinect_offset))
-            if 0 < int(loc2d[0]) < width and 0 < int(loc2d[1]) < height:
-                cv2.circle(color, (int(loc2d[0]), int(loc2d[1])), 1, (255, 255, 255), 1)
-        cv2.imwrite(os.path.join(output_dir, cam + f'Frame_.{frame_id}.jpg'), color)
+
+        # iterate through all meshes
+        for mesh in meshes:
+            points = np.asarray(mesh.vertices)
+            points2d = project_pose(points, params)
+            for loc2d in points2d:
+                # x, y = np.int16(np.round(loc2d - kinect_offset))
+                if 0 < int(loc2d[0]) < width and 0 < int(loc2d[1]) < height:
+                    cv2.circle(color, (int(loc2d[0]), int(loc2d[1])), 1, (255, 255, 255), 1)
+        # writes the image
+        cv2.imwrite(os.path.join(output_dir, f'Frame_.{frame_id}_{cam}.jpg'), color)
 
 
 # add a placeholder sphere
@@ -129,7 +134,7 @@ def add_mesh_sphere(point, vis, name):
 
 # display a 3D point in the world
 # the world consists of point clouds gathered from each camera
-def render_camera_poses(points, vis, frame_id):
+def render_camera_poses(vis, frame_id):
     # origin of the world coordinate system
     mesh_frame = o3d.geometry.TriangleMesh.create_coordinate_frame(size=2, origin=[0, 0, 0])
     vis.add_geometry("coordinate_frame", mesh_frame)
@@ -169,6 +174,16 @@ def render_camera_poses(points, vis, frame_id):
         vis.add_geometry(f"{cam}-depth-cam", depth_origin)
         vis.add_3d_label(depth_origin.get_center(), cam + 'depth-')
 
+
+
+    # ---- Printing 3D key points ----
+    # 3D - key points detected by voxel pose
+    with open('data/pred_voxelpose.pkl', 'rb') as f:
+        preds = pickle.load(f)
+
+    # 3d keypoints predicted with voxelpose of this frame
+    points = preds[frame_id]
+
     # will also print key points
     for i, joints in enumerate(points):
         joints = joints[:, :-2] / 1000
@@ -176,7 +191,7 @@ def render_camera_poses(points, vis, frame_id):
             add_mesh_sphere(point, vis, f"{coco_joints_def[j]}_Person: {i}")
 
 
-def render_smpl(frame_id, vis=None):
+def render_smpl(frame_id, vis):
     # translates the frame_id to index
     data_id = (frame_id - 2000) // 5
 
@@ -185,240 +200,145 @@ def render_smpl(frame_id, vis=None):
 
     # reads all smpl parameters
     data = read_smpl_all()
-    # for debugging purposes we only visualize one person
-    frame = data[data_id][1]
-    Rh = frame['Rh']
-    Th = frame['Th']
-    poses = frame['poses']
-    shapes = frame['shapes']
 
-    # gets the vertices
-    vertices = body_model(poses, shapes, Rh, Th, return_verts=True, return_tensor=False)[0]
-    # the mesh 
-    model = create_mesh(vertices=vertices, faces=body_model.faces)
+    meshes = []
+    # loads all meshes
+    for i in range(len(data[data_id])):
+        frame = data[data_id][i]
+        Rh = frame['Rh']
+        Th = frame['Th']
+        poses = frame['poses']
+        shapes = frame['shapes']
+
+        # gets the vertices
+        vertices = body_model(poses, shapes, Rh, Th, return_verts=True, return_tensor=False)[0]
+        # the mesh 
+        model = create_mesh(vertices=vertices, faces=body_model.faces)
+        meshes.append(model)
     
-    # projects vertices to 2d image
-    project_to_views(np.asarray(model.vertices), frame_id)
-    # project smpl onto 2d image with open3d render
-    project_smpl(frame_id, model)
-    # project smpl onto 2d image with pyrender
-    project_smpl_pyrender(frame_id, model)
+    
+    # projects vertices to 2D image
+    project_to_views(meshes, frame_id)
 
-    # project to 3D space
+    # projects SMPL mesh onto 2D image with pyrender
+    project_smpl_pyrender(meshes, frame_id)
+
+    # projects SMPL mesh to 3D space
     if vis != None:
-        for i in range(len(data[data_id])):
-            frame = data[data_id][i]
-            Rh = frame['Rh']
-            Th = frame['Th']
-            poses = frame['poses']
-            shapes = frame['shapes']
-            vertices = body_model(poses, shapes, Rh, Th, return_verts=True, return_tensor=False)[0]
-            model = create_mesh(vertices=vertices, faces=body_model.faces)
-            vis.add_geometry(f'Person {i}', model)
+        for i, mesh in enumerate(meshes):
+            vis.add_geometry(f'Person {i}', mesh)
 
 
-# WARNING: Bugged
-# rendering for camera 01 with pyrender
-# From: https://github.com/alon1samuel/Visualize3DHumanModels
-def project_smpl_pyrender(frame_id, mesh):
+# Renders meshes to the images
+def project_smpl_pyrender(meshes, frame_id):
     output_dir = prepare_out_dirs(dataDirs=['pyrender_on_img'])[0]
-    img = cv2.imread(os.path.join(DATA_DIR, 'cn01', '0000002150_color.jpg'), -1)
+
+    for cam in CAMERAS[:]:
+        file_id = str(frame_id).zfill(10)
+        params = load_camera_params(cam, DATA_DIR)
+        fpath = os.path.join(DATA_DIR, cam, f"{file_id}_color.jpg")
+        img = cv2.imread(fpath, cv2.IMREAD_UNCHANGED)
+        # shape is in form: [height, width, channel]
+        H, W, _ = img.shape
+        if img is None:
+            print("File not found: ", fpath)
+
+        fx = params['fx']
+        fy = params['fy']
+        cx = params['cx']
+        cy = params['cy']
+        R = params['R']
+        T = params['T']
+
+        # Our camera intrinsics matrix
+        pred_pose=np.eye(4)
+        pred_pose[:3, :3] = R
+        pred_pose[:3, 3] = T.flatten()
 
 
-    fx = 1.02595850e+03
-    fy = 1.02400476e+03
-    cx = 1.02028540e+03
-    cy = 7.76271240e+02
+        camera_pose = np.eye(4)
 
-    R = np.array(
-        [[-0.66923718,  0.74202016, -0.03908604],
-        [ 0.50586555,  0.4164543 , -0.75542432],
-        [-0.54426251, -0.52533031, -0.65406911]]
-    )
+        # Our Y goes down, in pyrender it goes up
+        camera_pose[1, 1] *= (-1.0)
 
-    T = np.array(
-        [[-0.22217233],
-        [-0.29406874],
-        [3.90767329]]
-    )
-  
-    vertices = np.asarray(mesh.vertices)
-    faces = np.asarray(mesh.triangles)
-
-    out_mesh = trimesh.Trimesh(vertices, faces)
-    rot = trimesh.transformations.rotation_matrix(
-        np.radians(180), [1, 0, 0])
-    out_mesh.apply_transform(rot)
-
-    # Create the camera object
-    H, W, _ = img.shape
-
-    baseColorFactor = (1.0, 1.0, 0.95, 1.0)
-
-    material = pyrender.MetallicRoughnessMaterial(
-        metallicFactor=0.0,
-        alphaMode='OPAQUE',
-        baseColorFactor=baseColorFactor)
-    mesh = pyrender.Mesh.from_trimesh(
-        out_mesh,
-        material=material)
-
-    scene = pyrender.Scene(bg_color=[0.0, 0.0, 0.0, 0.0],
-                           ambient_light=(0.3, 0.3, 0.3))
-    scene.add(mesh, 'mesh')
-
-    camera_center = np.array([cx, cy])
-    camera_transl = T.flatten()
-    # Equivalent to 180 degrees around the y-axis. Transforms the fit to
-    # OpenGL compatible coordinate system.
-    camera_transl[0] *= -1.0
-
-    camera_pose = np.eye(4)
-    # camera_pose[:3, :3] = R
-    camera_pose[:3, 3] = camera_transl
-
-    camera = pyrender.camera.IntrinsicsCamera(
-        fx=fx, fy=fy,
-        cx=camera_center[0], cy=camera_center[1])
-    scene.add(camera, pose=camera_pose)
-
-    # Get the lights from the viewer
-    # light_nodes = monitor.mv.viewer._create_raymond_lights()
-    # for node in light_nodes:
-    #     scene.add_node(node)
-
-    r = pyrender.OffscreenRenderer(viewport_width=W,
-                                   viewport_height=H,
-                                   point_size=1.0)
-
-    color, _ = r.render(scene, flags=pyrender.RenderFlags.RGBA)
-    color = color.astype(np.float32) / 255.0
-
-    valid_mask = (color[:, :, -1] > 0)[:, :, np.newaxis]
-
-    output_img = 255 - ((255 - color[:, :, :-1]) * valid_mask +
-                        (1 - valid_mask) * img)
-
-    output_img = (output_img * 255).astype(np.uint8)
-
-    cv2.imwrite(os.path.join(output_dir, f'Frame_{frame_id}.jpg'), output_img)
+        scene = pyrender.Scene(bg_color=[0.0, 0.0, 0.0, 0.0],
+                            ambient_light=(0.3, 0.3, 0.3))
 
 
+        camera = pyrender.camera.IntrinsicsCamera(
+            fx=fx, fy=fy,
+            cx=cx, cy=cy)
+
+        scene.add(camera, pose=camera_pose)
+
+        light = pyrender.PointLight(color=[1.0, 1.0, 1.0], intensity=8.0)
+        scene.add(light, pose=camera_pose)
+
+        # iterate through all meshes
+        for mesh in meshes:
+        
+            vertices = np.asarray(mesh.vertices)
+            faces = np.asarray(mesh.triangles)
+            out_mesh = trimesh.Trimesh(vertices, faces)
+
+            baseColorFactor = (1.0, 1.0, 0.95, 1.0)
+            material = pyrender.MetallicRoughnessMaterial(
+                metallicFactor=0.0,
+                alphaMode='OPAQUE',
+                baseColorFactor=baseColorFactor)
+            mesh = pyrender.Mesh.from_trimesh(
+                out_mesh,
+                material=material, smooth=True)
+
+            scene.add(mesh, 'mesh', pose=pred_pose)
+
+            # view the scene
+            # pyrender.Viewer(scene, use_raymond_lighting=True)
+
+        # Renderer of our scene
+        r = pyrender.OffscreenRenderer(viewport_width=W,
+                                    viewport_height=H,
+                                    point_size=1.0)
+
+        # Prints scene on our image
+        # From: https://github.com/alon1samuel/Visualize3DHumanModels/blob/bc531b40ac9eb67b3a7ff50139409a67b2a56adb/Demo.py#L145
+        color, _ = r.render(scene, flags=pyrender.RenderFlags.RGBA)
+        color = color.astype(np.float32) / 255.0
+
+        valid_mask = (color[:, :, -1] > 0)[:, :, np.newaxis]
+
+        output_img = 255 - ((255 - color[:, :, :-1]) * valid_mask +
+                            (1 - valid_mask) * img)
+
+        output_img = (output_img * 255).astype(np.uint8)
+
+        cv2.imwrite(os.path.join(output_dir, f'Frame_{frame_id}_{cam}.jpg'), output_img)
 
 
-# WARNING: Bugged
-# This is only implemented for camera 01 with open3d.visualization.rendering
-def project_smpl(frame_id, mesh):
-    output_dir = prepare_out_dirs(dataDirs=['open3d_render_on_img'])[0]
-    background_img = cv2.imread(os.path.join(DATA_DIR, 'cn01', '0000002150_color.jpg'), -1)
-
-    H, W, _ = background_img.shape
-
-    # create renderer
-    render = o3d.visualization.rendering.OffscreenRenderer(W, H)
-    # set background to black
-    render.scene.set_background([0.0, 0.0, 0.0, 1.0])
-
-
-    # Define a simple unlit Material.
-    # (The base color does not replace the arrows' own colors.)
-    mtl = o3d.visualization.rendering.MaterialRecord()
-    mtl.base_color = [1.0, 1.0, 1.0, 1.0]  # RGBA
-    mtl.shader = "defaultUnlit"
-
-    # adds our mesh
-    render.scene.add_geometry("mesh", mesh, mtl)
-
-    # camera intrinsics
-    camMat = np.array([
-                    [1.02595850e+03, 0., 1.02028540e+03],
-                    [0., 1.02400476e+03, 7.76271240e+02], 
-                    [0., 0., 1. ]])
-
-    # sets projection as in: https://stackoverflow.com/questions/70273002/project-3d-mesh-on-2d-image-using-camera-intrinsic-matrix
-    near_plane = 0.01
-    far_plane = 10.0
-    render.scene.camera.set_projection(camMat, near_plane, far_plane, W, H)
-
-    # Rotation matrix
-    R = np.array([
-        [-0.66923718,  0.74202016, -0.03908604],
-        [ 0.50586555,  0.4164543 , -0.75542432],
-        [-0.54426251, -0.52533031, -0.65406911]]
-    )
-
-    # Translation matrix
-    T = np.array([
-        [-0.22217233],
-        [-0.29406874],
-        [3.90767329]])
-    
-    # Position of our camera
-    # Taken from: https://math.stackexchange.com/questions/82602/how-to-find-camera-position-and-rotation-from-a-4x4-matrix
-    C = - R.T @ T
-
-    center = [0, 0, 0]  # look_at target
-    eye = C.flatten()  # camera position
-    up = [0, 0, 1]  # camera orientation
-    render.scene.camera.look_at(center, eye, up)
-
-
-
-    # Read the image into a variable
-    img_o3d = render.render_to_image()
-
-    # convert image to opencv
-    foreground_img = cv2.cvtColor(np.array(img_o3d), cv2.COLOR_RGBA2BGRA)
-
-    # making background transparent. taken from: 
-    # https://stackoverflow.com/questions/40527769/removing-black-background-and-make-transparent-from-grabcut-output-in-python-ope
-    tmp = cv2.cvtColor(foreground_img, cv2.COLOR_BGRA2GRAY)
-    _, alpha = cv2.threshold(tmp,0,255,cv2.THRESH_BINARY)
-    b, g, r, _ = cv2.split(foreground_img)
-    rgba = [b,g,r, alpha]
-    foreground_img = cv2.merge(rgba,4)
-    # cv2.imwrite("test.jpg", dst)
-
-    # merges two pictures, taken from: https://stackoverflow.com/questions/14063070/overlay-a-smaller-image-on-a-larger-image-python-opencv
-    y1, y2 = 0, foreground_img.shape[0]
-    x1, x2 = 0, foreground_img.shape[1]
-
-    alpha_s = foreground_img[:, :, 3] / 255.0
-    alpha_l = 1.0 - alpha_s
-
-    for c in range(0, 3):
-        background_img[y1:y2, x1:x2, c] = (alpha_s * foreground_img[:, :, c] +
-                                alpha_l * background_img[y1:y2, x1:x2, c])
-
-    cv2.imwrite(os.path.join(output_dir, f'Frame_{frame_id}.png'), background_img)
-
-
+# Set to True if you want to delete old output files
+DELETE_OLD_OUTPUT = False
 
 def main():
+    # Set this to True if you want to visualize the pointcloud
+    VISUALIZE_3D = True
     frame_id = 2150
     np.set_printoptions(suppress=True)
-    # 3D - key points detected by voxel pose
-    with open('data/pred_voxelpose.pkl', 'rb') as f:
-        preds = pickle.load(f)
 
-    vis = None
+    # renders smpl on 2d image
+    if not VISUALIZE_3D:
+        render_smpl(frame_id, None)
+    else:
+        # this will also visualize smpl model in 3d space
+        app = gui.Application.instance
+        app.initialize()
+        vis = o3d.visualization.O3DVisualizer("Open3D - 3D Text", 1024, 768)
+        vis.show_settings = True
+        render_camera_poses(vis, frame_id)
+        # Also visualizes SMPL Model in 3D Space
+        render_smpl(frame_id, vis)
 
-    # 3d keypoints predicted with voxelpose of this frame
-    points_ = preds[frame_id]
-
-    # this will also visualize smpl model in 3d space
-    app = gui.Application.instance
-    app.initialize()
-    vis = o3d.visualization.O3DVisualizer("Open3D - 3D Text", 1024, 768)
-    vis.show_settings = True
-    render_camera_poses(points_, vis, frame_id)
-
-    # renders smpl on 2d image and 3d space
-    # if vis != None then it will produce smpl in 3d point cloud
-    render_smpl(frame_id, vis)
-
-    app.add_window(vis)
-    app.run()
+        app.add_window(vis)
+        app.run()
 
 
 if __name__ == "__main__":
